+++ NEWS.md
@@ -1,5 +1,7 @@
 # samplr 1.1.0
 * ABS now can return estimates when using relative stopping rule.
+* The Bayesian Sampler function can now be used to not only return expected predictions but also expected variance
+* The Bayesian Sampler function can be now be used to return simulated predictions.
 
 # samplr 1.0.1
 * Fixes float comparisons using `!=` and thus not passing tests in some OS.